requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.11.0
sqlparse==0.5.5
